core: Create Premiere edition seam and per-clipitem rates
Split the Premiere generator into an edition shim (premiere.rb) and the
single-track core (premiere_core.rb) so a multi-track edition can load a
premiere_pro variant through ButterCut.engine_variant, which falls back
to core until that file exists.

Declare each clipitem's own <rate> so importers read in/out frames in
the source timebase — without it Premiere inherits the sequence rate
and misreads every trim on a rate-mismatched clip. Route clipitem
<enabled> through xmeml_enabled so an edition can disable a component;
nil (the usual case) stays TRUE, so core output is unchanged beyond the
new <rate> elements.
